+++ doc/changelog.rst
@@ -18,6 +18,8 @@ General
   introduction in 3.3.1 of :ref:`tendon armature<tendon-spatial-armature>`. In addition to the traditional
   ``mjData.qM``, :ref:`mj_makeM` also computes ``mjData.M``, a CSR representation of the same matrix.
 - Added a new function :ref:`mj_copyBack` to copy real-valued arrays in an mjModel to a compatible mjSpec.
+- Removed the limitation of :ref:`fusestatic<compiler-fusestatic>` to models which contain no references. The fusestatic
+  flag will now fuse all bodies which are not referenced and ignore bodies which are referenced.
 
 Simulate
 ^^^^^^^^